making derived model submodules 'private'
