+++ resqpy/derived_model/__init__.py
@@ -8,24 +8,24 @@ __all__ = [
     'zonal_grid', 'single_layer_grid', 'zone_layer_ranges_from_array'
 ]
 
-from .dm_add_edges_per_column_property_array import add_edges_per_column_property_array
-from .dm_add_faults import add_faults
-from .dm_add_one_blocked_well_property import add_one_blocked_well_property
-from .dm_add_one_grid_property_array import add_one_grid_property_array
-from .dm_add_single_cell_grid import add_single_cell_grid
-from .dm_add_wells_from_ascii_file import add_wells_from_ascii_file
-from .dm_add_zone_by_layer_property import add_zone_by_layer_property
-from .dm_coarsened_grid import coarsened_grid
-from .dm_copy_grid import copy_grid
-from .dm_drape_to_surface import drape_to_surface
-from .dm_extract_box_for_well import extract_box_for_well
-from .dm_extract_box import extract_box
-from .dm_fault_throw_scaling import fault_throw_scaling, global_fault_throw_scaling
-from .dm_gather_ensemble import gather_ensemble
-from .dm_interpolated_grid import interpolated_grid
-from .dm_local_depth_adjustment import local_depth_adjustment
-from .dm_refined_grid import refined_grid
-from .dm_tilted_grid import tilted_grid
-from .dm_unsplit_grid import unsplit_grid
-from .dm_zonal_grid import zonal_grid, single_layer_grid
-from .dm_zone_layer_ranges_from_array import zone_layer_ranges_from_array
+from ._dm_add_edges_per_column_property_array import add_edges_per_column_property_array
+from ._dm_add_faults import add_faults
+from ._dm_add_one_blocked_well_property import add_one_blocked_well_property
+from ._dm_add_one_grid_property_array import add_one_grid_property_array
+from ._dm_add_single_cell_grid import add_single_cell_grid
+from ._dm_add_wells_from_ascii_file import add_wells_from_ascii_file
+from ._dm_add_zone_by_layer_property import add_zone_by_layer_property
+from ._dm_coarsened_grid import coarsened_grid
+from ._dm_copy_grid import copy_grid
+from ._dm_drape_to_surface import drape_to_surface
+from ._dm_extract_box_for_well import extract_box_for_well
+from ._dm_extract_box import extract_box
+from ._dm_fault_throw_scaling import fault_throw_scaling, global_fault_throw_scaling
+from ._dm_gather_ensemble import gather_ensemble
+from ._dm_interpolated_grid import interpolated_grid
+from ._dm_local_depth_adjustment import local_depth_adjustment
+from ._dm_refined_grid import refined_grid
+from ._dm_tilted_grid import tilted_grid
+from ._dm_unsplit_grid import unsplit_grid
+from ._dm_zonal_grid import zonal_grid, single_layer_grid
+from ._dm_zone_layer_ranges_from_array import zone_layer_ranges_from_array
