+++ resqpy/derived_model/_dm_drape_to_surface.py
@@ -13,8 +13,8 @@ import resqpy.model as rq
 import resqpy.olio.intersection as meet
 import resqpy.olio.xml_et as rqet
 
-from resqpy.derived_model.dm_common import __displacement_properties, __prepare_simple_inheritance, __write_grid, __establish_model_and_source_grid
-from resqpy.derived_model.dm_copy_grid import copy_grid
+from resqpy.derived_model._dm_common import __displacement_properties, __prepare_simple_inheritance, __write_grid, __establish_model_and_source_grid
+from resqpy.derived_model._dm_copy_grid import copy_grid
 
 
 def drape_to_surface(epc_file,
