+++ resqpy/derived_model/_dm_fault_throw_scaling.py
@@ -13,8 +13,8 @@ import resqpy.model as rq
 import resqpy.olio.uuid as bu
 import resqpy.olio.xml_et as rqet
 
-from resqpy.derived_model.dm_common import __displacement_properties, __prepare_simple_inheritance, __write_grid, __establish_model_and_source_grid
-from resqpy.derived_model.dm_copy_grid import copy_grid
+from resqpy.derived_model._dm_common import __displacement_properties, __prepare_simple_inheritance, __write_grid, __establish_model_and_source_grid
+from resqpy.derived_model._dm_copy_grid import copy_grid
 
 
 def fault_throw_scaling(epc_file,
