+++ resqpy/derived_model/_dm_tilted_grid.py
@@ -1,4 +1,4 @@
-"""High level unsplit grid function."""
+"""High level tilted grid function."""
 
 import logging
 
@@ -7,25 +7,35 @@ log = logging.getLogger(__name__)
 import os
 
 import resqpy.model as rq
+import resqpy.olio.vector_utilities as vec
 import resqpy.olio.xml_et as rqet
 
-from resqpy.derived_model.dm_common import __prepare_simple_inheritance, __write_grid, __establish_model_and_source_grid
-from resqpy.derived_model.dm_copy_grid import copy_grid
+from resqpy.derived_model._dm_common import __displacement_properties, __prepare_simple_inheritance, __write_grid, __establish_model_and_source_grid
+from resqpy.derived_model._dm_copy_grid import copy_grid
 
 
-def unsplit_grid(epc_file,
-                 source_grid = None,
-                 inherit_properties = False,
-                 inherit_realization = None,
-                 inherit_all_realizations = False,
-                 new_grid_title = None,
-                 new_epc_file = None):
-    """Extends epc file with a new grid which is a version of the source grid with all faults healed.
+def tilted_grid(epc_file,
+                source_grid = None,
+                pivot_xyz = None,
+                azimuth = None,
+                dip = None,
+                store_displacement = False,
+                inherit_properties = False,
+                inherit_realization = None,
+                inherit_all_realizations = False,
+                new_grid_title = None,
+                new_epc_file = None):
+    """Extends epc file with a new grid which is a version of the source grid tilted.
 
     arguments:
        epc_file (string): file name to rewrite the model's xml to; if source grid is None, model is loaded from this file
        source_grid (grid.Grid object, optional): if None, the epc_file is loaded and it should contain one ijk grid object
           (or one 'ROOT' grid) which is used as the source grid
+       pivot_xyz (triple float): a point in 3D space on the pivot axis, which is horizontal and orthogonal to azimuth
+       azimuth: the direction of tilt (orthogonal to tilt axis), as a compass bearing in degrees
+       dip: the angle to tilt the grid by, in degrees; a positive value tilts points in direction azimuth downwards (needs checking!)
+       store_displacement (boolean, default False): if True, 3 grid property parts are created, one each for x, y, & z
+          displacement of cells' centres brought about by the tilting
        inherit_properties (boolean, default False): if True, the new grid will have a copy of any properties associated
           with the source grid
        inherit_realization (int, optional): realization number for which properties will be inherited; ignored if
@@ -35,14 +45,10 @@ def unsplit_grid(epc_file,
           inherit_properties is False or inherit_realization is not None
        new_grid_title (string): used as the citation title text for the new grid object
        new_epc_file (string, optional): if None, the source epc_file is extended with the new grid object; if present,
-          a new epc file (& associated h5 file) is created to contain the unsplit grid (& crs)
+          a new epc file (& associated h5 file) is created to contain the tilted grid (& crs)
 
     returns:
-       a new grid (grid.Grid object) which is an unfaulted copy of the source grid
-
-    notes:
-       the faults are healed by shifting the thrown sides up and down to the midpoint, only along the line of the fault;
-       to smooth the adjustments away from the line of the fault, use the global_fault_throw_scaling() function first
+       a new grid (grid.Grid object) which is a copy of the source grid tilted in 3D space
     """
 
     assert epc_file or new_epc_file, 'epc file name not specified'
@@ -54,35 +60,35 @@ def unsplit_grid(epc_file,
     assert source_grid.grid_representation == 'IjkGrid'
     assert model is not None
 
-    assert source_grid.has_split_coordinate_lines, 'source grid is unfaulted'
-
     # take a copy of the grid
     grid = copy_grid(source_grid, model)
 
     if grid.inactive is not None:
         log.debug('copied grid inactive shape: ' + str(grid.inactive.shape))
 
-    # heal faults in the grid
+    # tilt the grid
     grid.cache_all_geometry_arrays()  # probably already cached anyway
-    unsplit = source_grid.unsplit_points_ref()
-    grid.points_cached = unsplit.copy()
-    assert grid.points_cached.shape == (grid.nk + 1, grid.nj + 1, grid.ni + 1, 3), 'unsplit points have incorrect shape'
+    vec.tilt_points(pivot_xyz, azimuth, dip, grid.points_cached)
 
-    grid.has_split_coordinate_lines = False
-    delattr(grid, 'split_pillar_indices_cached')
-    delattr(grid, 'cols_for_split_pillars')
-    delattr(grid, 'cols_for_split_pillars_cl')
-    if hasattr(grid, 'pillars_for_column'):
-        delattr(grid, 'pillars_for_column')
+    # build cell displacement property array(s)
+    if store_displacement:
+        displacement_collection = __displacement_properties(grid, source_grid)
+    else:
+        displacement_collection = None
 
     collection = __prepare_simple_inheritance(grid, source_grid, inherit_properties, inherit_realization,
                                               inherit_all_realizations)
-    # todo: recompute depth properties (and volumes, cell lengths etc. if being strict)
+    if collection is None:
+        collection = displacement_collection
+    elif displacement_collection is not None:
+        collection.inherit_imported_list_from_other_collection(displacement_collection, copy_cached_arrays = False)
 
     if new_grid_title is None or len(new_grid_title) == 0:
-        new_grid_title = 'unfaulted version of ' + str(rqet.citation_title_for_node(source_grid.root))
+        new_grid_title = 'tilted version ({0:4.2f} degree dip) of '.format(abs(dip)) + str(
+            rqet.citation_title_for_node(source_grid.root))
 
     # write model
+    model.h5_release()
     if new_epc_file:
         __write_grid(new_epc_file, grid, property_collection = collection, grid_title = new_grid_title, mode = 'w')
     else:
